task filter for priorities, fix bugs and refactoring

+++ app/src/main/java/com/manoelh/task/entity/PriorityEntity.kt
@@ -1,6 +1,6 @@
 package com.manoelh.task.entity
 
-data class PriorityEntity (var id: Long, var description: String){
+data class PriorityEntity (var id: Int, var description: String){
 
     override fun toString(): String {
         return description
